Refactor SSL implementations: * Using connection_ssl.hpp for both connection_tls/_ssl source files * Renamed USING_OPENSSL to USE_OPENSSL * Move all specific variables to the SSLContext struct * init_ssl changed to a standalone function * Added tls_key_logging config to connection_ssl * Seperated logging functions into separate ssl and tls files * Added default_delete in connection_tls * Using openssl SSL_set_ex_data/get_ex_data function to get the udp tls key logging server in the openssl keylog callback * Using std::invoke
