+++ 3rd_party/CMakeLists.txt
@@ -1,3 +1,3 @@
-if ((NOT USING_OPENSSL) AND ISO15118_LINK_CUSTOM_MBEDTLS)
+if ((NOT USE_OPENSSL) AND ISO15118_LINK_CUSTOM_MBEDTLS)
     add_subdirectory(mbedtls)
 endif ()
