+++ include/iso15118/detail/helper.hpp
@@ -7,6 +7,7 @@
 #include <cstdarg>
 #include <cstdint>
 #include <cstdio>
+#include <functional>
 #include <string>
 
 namespace iso15118 {
@@ -27,10 +28,6 @@ void log(const LogLevel&, const std::string&);
 
 void log_and_throw(const char* msg);
 
-void log_and_raise_mbed_error(const std::string& error_msg, int error_code);
-
-void log_and_raise_openssl_error(const std::string& error_msg);
-
 std::string adding_err_msg(const std::string& msg);
 
 template <typename CallbackType, typename... Args> bool call_if_available(const CallbackType& callback, Args... args) {
@@ -38,7 +35,7 @@ template <typename CallbackType, typename... Args> bool call_if_available(const 
         return false;
     }
 
-    callback(std::forward<Args>(args)...);
+    std::invoke(callback, std::forward<Args>(args)...);
     return true;
 }
