+++ include/iso15118/io/connection_ssl.hpp
@@ -14,12 +14,10 @@ namespace iso15118::io {
 struct SSLContext;
 class ConnectionSSL : public IConnection {
 public:
-    ConnectionSSL(PollManager&, const std::string&, const config::SSLConfig&);
+    ConnectionSSL(PollManager&, const std::string& interface_name, const config::SSLConfig&);
 
     void set_event_callback(const ConnectionEventCallback&) final;
     Ipv6EndPoint get_public_endpoint() const final;
-
-    void init_ssl(const config::SSLConfig&);
 
     void write(const uint8_t* buf, size_t len) final;
     ReadResult read(uint8_t* buf, size_t len) final;
@@ -34,17 +32,9 @@ private:
 
     Ipv6EndPoint end_point;
 
-    int fd{-1};
-    int accept_fd{-1};
-
-    bool connection_open{false};
-
     ConnectionEventCallback event_callback{nullptr};
 
     bool handshake_complete{false};
-
-    std::string interface_name;
-    bool enable_ssl_logging{false};
 
     void handle_connect();
     void handle_data();
