+++ include/iso15118/io/sdp_server.hpp
@@ -2,6 +2,9 @@
 // Copyright 2023 Pionix GmbH and Contributors to EVerest
 #pragma once
 
+#include <cstdint>
+#include <string>
+
 #include <netinet/in.h>
 
 #include "ipv6_endpoint.hpp"
@@ -37,6 +40,23 @@ public:
 private:
     int fd{-1};
     uint8_t udp_buffer[2048];
+};
+
+class TlsKeyLoggingServer {
+public:
+    TlsKeyLoggingServer(const std::string& interface_name, uint16_t port);
+    ~TlsKeyLoggingServer();
+
+    ssize_t send(const char* line);
+
+    auto get_fd() const {
+        return fd;
+    }
+
+private:
+    int fd{-1};
+    uint8_t buffer[2048];
+    sockaddr_in6 destination_address{};
 };
 
 } // namespace iso15118::io
