+++ src/iso15118/CMakeLists.txt
@@ -79,7 +79,7 @@ target_link_libraries(iso15118
         Threads::Threads
 )
 
-if (USING_OPENSSL)
+if (USE_OPENSSL)
     target_compile_definitions(iso15118
         PRIVATE
             LIBISO_OPENSSL
@@ -91,7 +91,8 @@ if (USING_OPENSSL)
     )
     target_sources(iso15118
         PRIVATE
-        io/connection_ssl.cpp
+        io/connection_openssl.cpp
+        misc/helper_openssl.cpp
     )
 else()
     if (NOT ISO15118_LINK_CUSTOM_MBEDTLS)
@@ -113,7 +114,8 @@ else()
     )
     target_sources(iso15118
         PRIVATE
-        io/connection_tls.cpp
+        io/connection_mbedtls.cpp
+        misc/helper_mbedtls.cpp
     )
 endif()
